plugins: single source of truth
Change the plugin interface to collect nix attribute names and database
names from the ResourceState record instead of the XBackendRegistration.

The idea is to simplify the code paths and have a clearer, simpler
place for plugins' definitions and states to meet -- but it is getting
away from the goal of making nixops-aws work this week.